Feature/update for new rust UI (#29)
* Updated Color Preview coordinates

+++ src/main/java/com/bobrust/util/debug/DebugUtil.java
@@ -9,11 +9,11 @@ public class DebugUtil {
 	}
 	
 	@Deprecated
-	public static void debugShowImage(Image image, int scale) {
-		int width = image.getWidth(null) * scale;
-		int height = image.getHeight(null) * scale;
+	public static void debugShowImage(Image image, double scale) {
+		double width = image.getWidth(null) * scale;
+		double height = image.getHeight(null) * scale;
 		
-		Image scaled = image.getScaledInstance(width, height, Image.SCALE_FAST);
+		Image scaled = image.getScaledInstance((int) width, (int) height, Image.SCALE_FAST);
 		ImageIcon icon = new ImageIcon(scaled);
 		JOptionPane.showMessageDialog(null, new JLabel(icon), "Debug image", JOptionPane.INFORMATION_MESSAGE);
 	}
